requirements.txt
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
attrs==25.3.0
Automat==25.4.16
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
constantly==23.10.4
cryptography==45.0.7
cssselect==1.3.0
defusedxml==0.7.1
distro==1.9.0
filelock==3.19.1
Flask==3.1.2
groq==0.31.1
h11==0.16.0
http-exceptions==0.2.10
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.2
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
lxml==6.0.1
MarkupSafe==3.0.2
openai==1.107.2
packaging==25.0
parsel==1.10.0
pillow==11.3.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
python-dotenv==1.1.1
queuelib==1.8.0
requests==2.32.5
requests-file==2.1.0
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
tldextract==5.3.0
toolhouse==1.4.0
tqdm==4.67.1
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
w3lib==2.3.1
Werkzeug==3.1.3
zope.interface==8.0
